Triton Renaming and Add bert example for triton inference server (#750)
* Add bert triton example

* fix init

* Triton reference

* Rename to triton

* Add bert transformer to e2e image build

* Generate for triton sdk types

* Add to bert transformer build to CI

* Add e2e test for triton

* Add back license

* Add gpu annotation on test

* Fix storage uri in the test

* upgrade test cluster

* Add debug for the pod

* Fix test resource

* Add retry

* Test bert model with triton

* Skip triton test

* Skip tag resolution for nvcr.io

* Use simpler example

* skip tag resolution for nvcr.io

* Upgrade to Knative 0.15.0 for e2e testing

* Upgrade to kubernetes 1.16

* Add note for knative version dependency

+++ hack/python-sdk/client-gen.sh
@@ -26,9 +26,19 @@ SWAGGER_CODEGEN_FILE="pkg/apis/serving/v1alpha2/swagger.json"
 SDK_OUTPUT_PATH="python/kfserving"
 
 echo "Downloading the swagger-codegen JAR package ..."
-wget -O ${SWAGGER_CODEGEN_JAR} ${SWAGGER_JAR_URL}
+if [ ! -f ${SWAGGER_CODEGEN_JAR} ]
+then
+    wget -O ${SWAGGER_CODEGEN_JAR} ${SWAGGER_JAR_URL}
+fi
 
 echo "Generating Python SDK for KFServing ..."
 java -jar ${SWAGGER_CODEGEN_JAR} generate -i ${SWAGGER_CODEGEN_FILE} -l python -o ${SDK_OUTPUT_PATH} -c ${SWAGGER_CODEGEN_CONF}
 
+# revert following files since they are diveraged from generated ones
+git checkout python/kfserving/kfserving/rest.py
+git checkout python/kfserving/README.md
+git checkout python/kfserving/kfserving/__init__.py
+git checkout python/kfserving/test/__init__.py
+git checkout python/kfserving/setup.py
+git checkout python/kfserving/requirements.txt
 echo "KFServing Python SDK is generated successfully to folder ${SDK_OUTPUT_PATH}/."
